Restrict maximum urllib3 version for requests compatibility
Previously, no references to the urllib3 package were present in the
setup.py file. The latest version of requests (2.19.1), a transitive
dependency brought in from dxlclient, disallows versions less than 1.24
of urllib3 from being installed. The latest version of the
elasticsearch 5.x dependency (5.5.5), however, does not place a
restriction on the maximum version of urllib3.

A pip install now pulls in urllib3 version 1.24, which was just
released. This causes an install error due to the maximum urllib3
version restriction from the requests package.

For compatibility with the requests library, this commit adds a
<1.24 maximum version restriction on urllib3 in order to allow pip
installations to work properly again.